add lgamma to op_threshold_white_list
