+++ python/paddle/fluid/tests/unittests/white_list/op_threshold_white_list.py
@@ -45,6 +45,7 @@ NEED_FIX_FP64_CHECK_GRAD_THRESHOLD_OP_LIST = [
     'bilateral_slice',\
     'cudnn_lstm', \
     'rnn', \
+    'lgamma', \
 ]
 
 NEED_FIX_FP64_CHECK_OUTPUT_THRESHOLD_OP_LIST = ['bilinear_interp',\
